pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "spectral"         # make sure your code lives in src/spectral/
version = "0.1.0"
description = "Advanced Numerical Algorithms – codebase"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "Philip Korsager Nickel" }]

# All dependencies (runtime, docs, and dev)
dependencies = [
  # Runtime dependencies
  "dtumathtools>=2025.2.0",
  "numba>=0.62.1",
  "opencv-python>=4.11.0.86",
  "pandas>=2.3.3",
  "pyarrow>=22.0.0",
  "scipy>=1.16.2",
  "seaborn>=0.13.2",
  # Documentation dependencies
  "sphinx>=8.2.3",
  "pydata-sphinx-theme>=0.15.0",
  "numpydoc>=1.8.0",
  "sphinx-copybutton>=0.5.2",
  "sphinx-gallery>=0.17.0",
  "pillow>=10.0.0", # Required for sphinx-gallery image handling
  # Development dependencies
  "ruff>=0.6.9",
  "pytest>=8.3.3",
  "pytest-cov>=5.0.0",
  "interrogate>=1.7.0",
  "pydocstyle>=6.3.0",
  "line-profiler>=5.0.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

# include = ["mainrepo*"]   # optionally narrow discovery

[tool.setuptools.package-data]
spectral = ["styles/*.mplstyle"]

[dependency-groups]
dev = [
    "memray>=1.19.1",
    "py-spy>=0.4.1",
    "snakeviz>=2.2.2",
]
